Moved a few definitions closer to their use
