+++ sql/expression/function/connection_id_test.go
@@ -15,7 +15,7 @@ func TestConnectionID(t *testing.T) {
 	session := sql.NewSession("", "", "", 2)
 	ctx := sql.NewContext(context.Background(), sql.WithSession(session))
 
-	connIDFunc := sql.NewFunction0("connection_id", sql.Uint32, connIDFuncLogic)
+	connIDFunc := sql.NewFunction0("connection_id", NewConnectionID)
 	result, err := connIDFunc.Fn().Eval(ctx, nil)
 	require.NoError(err)
 	require.Equal(uint32(2), result)
